add multiple versions of same package capability and fix version param
namevar is required to be unique across package resources.  puppet does
not provide an alias for the package name and therefore uses namevar.
if you wish to install multiple versions of the same package, simply
set the namevar to pkgname<space>version and the version will be
ignored except from a namespace perspective.   Assigning the version
to the ensure => will be the actual version passed to chocolatey.
This also now functional so that it appends -version switch
if ensure is a defined string.