Fix for new version of chocolatey
It looks like the most recent version of chocolatey is giving us different output on the calls
this provider makes. This patch will parse the new format.

I removed one test due to it being irrelvant (We aren't going to receive bad output - we know
the format we are going to be getting from chocolatey.) and patched a few others as these are checking
implementation over behavior - So the changes made of course broke the tests.