(maint) use existing puppet/facter/hiera
If the Environment variables are set for puppet/facter/hiera,
use them instead of installing the gems for those. This is helpful
for those that work on those code bases and don't want to
constantly have to struggle with deleting facter gem when working
on the code base (facter is a special snowflake when it comes to
looking for facts along the loadpath and makes having it installed
impossible to update the code reliably).