(GH-29) Add additional default installation path for chocolatey
chocolatey 0.9.8.24 comes with new installation path -
C:\ProgramData\chocolatey. When you install this version of chocolatey
and use the provider in the same puppet run you will get an error
message saying that the provider is not functional on this host.

This change addresses the issue by checking if the new installation path
is present. If it is not it defaults to the new installation path. The check for legacy path is to
avoid breaking the provider for hosts running older versions of chocolatey.