(PUP-1691) Add custom fact for choco install path
This includes a custom fact for chocolatey's installpath. It will attempt
to read the environment variable for ChocolateyInstall from the registry
and default to 'C:\ProgramData\Chocolatey' if it is not found. This allows
Chocolatey to look at an existing installation by defaut.

The `$chocolatey::params::install_location` variable will default to
this custom fact, but can be overridden when declaring the class.