(GH-66) Comprehensive check choco.exe location
The current check only looks for the existence of directories with
fallback to a default location if it doesn't find any. The first check
against the environment variable may be incorrect if an old environment
variable is still lingering and can cause issues in finding the proper
provider location.

Check in a very explicit order for the existence of the choco.exe binary,
starting with default install, then the environment variable, then the old
default location, followed by the last check in the AllUsersProfile location
if nothing else can be found.