(GH-71) Use upgrade instead of install
When ensure => $version, the underlying package provider will call
install instead of upgrade if the package is already installed.
Chocolatey's default install behavior is to do nothing if a package is
already installed, even if there is a newer version available. In most
cases we can call upgrade, except when the installer is using a
packages.config.

We'll also need to address the use of side by side installs, which this
will remove the ability to do.