(PUP-1691) Install Chocolatey
Allow installing chocolatey with a custom package location. Also allow
using the built-in windows shell to unzip the package. The original request for
this had the ability to use the built-in Windows shell to perform the unzipping
and it keeps folks from having to download an external executable by
default. Allow 25 minutes for it to finish.