(PUP-1691) Ensure environment variables updated
Ensure that the windows environment variables are set and are present.
This should also allow Chocolatey to be installed and used on the first
run of Puppet as it will refresh the PATH environment variable so that
suitability is redetermined[1]. This stems from an original conversation[2] with
a few community members and @joshcooper about suitability determination not
being cached in newer versions of Puppet.

[1] https://docs.puppetlabs.com/guides/provider_development.html#suitability
[2] https://groups.google.com/d/topic/chocolatey/0hPNtAMmhNM/discussion