(GH-86) avoid using stale version of chocolateyinstall env var
In a case where chocolatey was recently installed but the shell
was not refreshed, the exec resource could be run with a stale
path (ie C:\chocolatey) if a migration to C:\ProgramData\chocolatey
had recently occured.

In this scenario, a folder structure will be improperly created
in the legacy path.

This patch ensures that the exec uses the puppet specified value
for the install location.