(MODULES-3138) Fix Reference Test Pre-suite
The reference tests pre-suite was using the "agents" host array when
calling the "install_dev_puppet_module_on". However, a method used within
the "install_dev_puppet_module_on" did not support an array of hosts.
The fix is to feed "install_dev_puppet_module_on" a single host object in a
block.