(GH-100) Do not call choco with --debug --verbose
The output from error messages is way too much when these are on by
default. It would be helpful to see only what has errored and allow the
user to investigate further if need be. This gives the user the ability
to choose whether or not to add those values to the arguments as part
of `install_options` and `uninstall_options`.