(maint) Remove choco location from verification
With the tests that are validating arguments passed to choco, we were
also verifying the path to Chocolatey, which is subject to influence by
the system and may actually differ based on actual path versus what the
registry Environment variable states. Since we only care about
validating the correct arguments are passed to Chocolatey, where the
choco executable is located is not a concern for the tests. Remove it
from the verification of the specs.