(MODULES-3430) Chocolateysource: Explicitly require location
When ensuring a chocolateysource is present, location is a required value
and should be explicitly set. Previously, location would just take on the
default value of the resource title. This could be alarming to some users
who may set a resource like:

~~~puppet
chocolateysource {'chocolatey':}
~~~

And be later surprised to find that it replaces the location value with
`chocolatey` instead of whatever it was originally, in this case that
value was `https://chocolatey.org/api/v2/`. Due to this, which could be
both alarming and upsetting, the resource should instead explicitly
require a location when `ensure => present`. When disabling or
removing a chocolateysource, it should be fine to simply name the
source to disable/remove.