(GH-125) Wrap PowerShell Redirect in Try/Catch
If the PowerShell Output redirection fix does not successfully apply,
we don't want it to fail the entire script. It uses internal variables
that only apply to the PowerShell InternalHost/ConsoleHost that
short-circuited logging and caused a FileStream error when you
attempted to write to multiple log streams.

Due to sometimes calling this script with non-default PowerShell
hosts (like with the new 2.0 puppet powershell provider), we want to
ensure that if the fix errors, it does not fail the entire script.