(maint) Remove travis and metadata-json-lint gems
To allow for specs to run correctly on older versions of Ruby, json
requires using less than 2.0.0. Interestingly though, json should not
be used for regular testing as it requires devkit. Remove both gems
that depend on json so that tests in older versions of Windows will
work and so that testing doesn't require DevKit.