(MODULES-3391) Limit `rake spec` on older puppet
The default `rake spec` task provided by puppetlabs_spec_helper requires
symlink support in puppet to put fixtures in place for testing. Older versions
of Puppet do not support symlinks on Windows. This commit adds a check in the
Rakefile for whether or not Puppet supports symlinks, and if it does not,
redefines the `rake spec` task that has been defined by puppetlabs_spec_helper,
to accomplish two objectives:
- Remove the call to `rake spec_prep` which prepares all fixtures and requires
  symlinks
- Restrict the pattern for matching tests to those under the 'unit' and
  'integration' directories (other tests require the fixture we won't have)

This feels a bit like overreaching into puppetlabs_spec_helper, but the
alternative seems to be to duplicate the `rake spec` task in its entirety here
or make even greater assumptions in puppetlabs_spec_helper about the
constitution of modules other than puppetlabs-chocolatey.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>