(MODULES-3758) Fix: chocolateysource user ensure sync
During the check for insync, user keeps getting detected for synchronize
due to the default value being set to '' as compared to a nil value
being returned from the existing items.

This can cause errors as it could set a source to enabled, even if the
source should be disabled. Instead we should set a default for the
existing item so it matches the default in the user property.