(MODULES-3430) Validate - Location Not Empty
When validating at the type level, typically all parameters/properties
has been set. The only instance of this not being true is with `puppet
resource`. With puppet resource, values are initially set to their
defaults when validation is called. This means location is sometimes
empty coming from the resource path. When validation runs, the resource
creation has not actually set the properties from the property_hash.
This causes the top-level resource validation to fail.

The resource is created and passed to provider.initialize,
which sets the property_hash differently than when run through the
normal manifest. https://github.com/puppetlabs/puppet/blob/5bb93f0c06cca6cd16f693e9dbaf2066d43a7b6a/lib/puppet/provider.rb#L499-L519

This is due to the call to self.instances resulting in initializing the
instances prior to setting properties - see
https://github.com/puppetlabs/puppet/blob/5bb93f0c06cca6cd16f693e9dbaf2066d43a7b6a/lib/puppet/type.rb#L1179-L1180.
Then it calls initialize still not having set the properties, it runs
validate -
https://github.com/puppetlabs/puppet/blob/5bb93f0c06cca6cd16f693e9dbaf2066d43a7b6a/lib/puppet/type.rb#L2391

This only happens during `puppet resource type` - when running Puppet
normally, it does not run into this issue. This seems to be a small
issue in how puppet resource works with self.instances and validation
at the type level.

Look in multiple places to find the value of `location`.