(MODULES-3430) Flush - Ensure / Disable Existing
When flushing values, be sure the value of ensure is found. It may not
be part of what is changing for the flush - it could be another
property. It is imperative to know the value of the ensure, so start
with `property_hash[:ensure]` value, then look at
`property_flush[:ensure]` if it is available.
If ensure is still not found, look at `resource[:ensure]`.

When disabling a source, that source must first exist. Disabling a
source removes two calls to choco, one on the version and another for
adding the source prior to disabling. Unfortunately it doesn't easily
allow bringing on a new machine with the same manifests when wanting to
disable a source that hasn't priorly existed. This is a very slight
edge case. Typically few sources would be disabled, most would
outright be removed.