(MODULES-4135) choco -v - Remove all extraneous messaging
When determining the version of Chocolatey that is installed, the
module calls `choco -v` which will return the version in most versions
of Chocolatey (anything earlier than 0.9.8.24 or more than 4 years old
may not). However sometimes there is messaging that tends to be
presented as warnings that still end up being included in the output.
Instead of looking for specific wording, typically the chocolatey
version is going to be the last line in the output. So remove all of
the other lines and just use the last line available.

Using a simpler `string.lines.last` is out because it only splits
on `$/` and doesn't allow multiple options. So split on CRLF, then LF,
then CR. This allows the output to be evaluated on different
platforms and not be subject to whatever Ruby believes `$/` evaluates
to on that platform.