(maint) Ensure versioncmp has a value when undef
When looking at determining the version of Chocolatey for
configuration purposes in config.pp, determine if the value of
`$chocolatey::chocolatey_version` is coming back undefined. If
it is, return a string with 0 as the value. Add a local variable
to make that determination and use it instead of the fact in the
versioncmp checks.