(MODULES-4091) Explicitly Set ChocolateyInstall Environment
Previously, ensuring the installation directory for Chocolatey
keyed off of an environment value passed to an exec powershell
resource through the environment parameter. However there is a
bug in the powershell module in versions 2.x up to 2.1.0
(current version as of this commit) that ignore the environment
parameter. Since it is better to explicitly set that value, take
a dependency on the registry module and explicitly ensure the
environment variable is set so that Chocolatey will successfully
install to an alternative install location.