(MODULES-4149) Unsuitable providers should not error
When attempting to force the provider, validation would error due
to the version coming back would not be above the version that is
being validated. When the provider is unsuitable, it should not
return an error during the Puppet run and should instead just
return the unsuitable message.

Check to ensure Chocolatey is installed prior to allowing the
validation check to not fail the Puppet run.