(MODULES-4056) Helpful Error if No Sources Enabled
Previously, when there were no sources enabled, Puppet would
attempt to loop over a non-existent set of packages pulled back
from listing a set of packages locally. Unfortunately this is due
to an issue in Chocolatey that requires at least one source to be
enabled to produce the local set of installed packages:
https://github.com/chocolatey/choco/issues/661

Instead of allowing Puppet to fail on a nil:NilClass error, provide
a helpful and actionable message instead.