(maint) Change InstallChocolatey ERB to be LF line endings
Previously in commit d102498b98, the git attributes file was modified to enforce
all ERB files have LF line endings.  However the InstallChocolatey.ps.erb file
had already been checked in with CRLF.  This caused the file to always be
marked as changed when checked out on Windows.  This commit changes the line
ending to be LF instead of CRLF, to be inline with .gitattributes.