(MAINT) Fix .gitattributes for module development on Windows
Previously the .gitattributes file, which is managed by modulesync, was updated
to set ERB files to LF line endings.  However on Windows ERB files can be used
to generate files in either line endings.  Also as the current ERB file is CRLF
when this file is cloned on a Window filesystem, it detects this file has
changed, which in turn blocks git functions such as rebasing or checking out.

This commit modifies the .sync.yml to exclude ERB files from the .gitattributes
setting and let these files manage line endings on their own.