(MODULES-4678) Explicitly close config on read
Previously, any providers managing Chocolatey configuration are unable
to be used with 0.10.4+ and actually cause the process to
fail.

Chocolatey 0.10.4+ has changed the way it updates and accesses the
config. The new method has exposed an error in an unclosed file handle
when reading the configuration for `chocolateysource`,
`chocolateyfeature`, and `chocolateyconfig`. Switch to `File.read` as
it will automatically close the file after reading out the contents.