(MODULES-4562) Use actual choco.exe
Sometimes when running the shimmed choco, it can cause issues and
it ends up returning the version of choco as '0'. When managing
configuration, this can cause issues using those resources, such
as the following:

`Failed to apply catalog: Validation of Chocolateysource[chocolatey]
failed: Chocolatey version must be '0.9.9.0' to manage configuration
values with Puppet. Detected '0' as your version. Please upgrade
Chocolatey to use this resource.`

On the next run of Puppet, everything works again appropriately. The
event log shows an issue with attempting to run the shim. In order
to avoid this altogether, just use choco.exe directly when running
Puppet.