(MODULES-5383) Remove exit codes arg with feature
When the "usepackageexitcodes" feature is enabled with the
chocolateyfeature resource, this commit ensures that the
--ignore-package-exit-codes argument is not passed to the choco binary
and that the feature is respected.

Previously, the --ignore-package-exit-codes argument would be passed all
the time (on versions of Chocolatey where it existed) and this would
result in the enabling of the "usepackageexitcodes" feature to be
rendered moot.