(MODULES-5383) Ensure feature was explicitly set.
Chocolatey enables the "usepackageexitcodes" by default, therefore this
will break existing behaviour of puppetlabs-chocolatey if it we do not
check that the feature was set explicitly.

This commit adds new methods to the provider to retrieve chocolatey
features from the configuration file. It uses these methods to
verify whether or not the `usepackageexitcodes` configuration setting
is explicitly set to enabled. If it is disabled or implicitly set,
the provider will add the `--ignore-package-exit-codes` parameter when
installing packages as is the existing behavior. If it is explicitly
set to enabled then the provider will *not* pass that parameter.

This is likely to break runs but respects the configuration settings.