(MODULES-6652) Make 7zip DL source configurable
Prior to this commit the 7zip executable, a requirement for the
installation of Chocolatey, was hardcoded for download from a
single internet URL. This caused failures when pointed through
a proxy.

This commit adds a new parameter, `$seven_zip_download_url` to
the Chocolatey class. This is used to specify the path to the
`7za.exe` binary, which can live anywhere that the Puppet File
resource can retrieve it for the target machine. It uses such
a file declaration in the `install.pp` manifest to handle cases
where the use of 7zip is required. This parameter defaults to
the previously hardcoded URL, `https://chocolatey.org/7za.exe`.

This commit also adds a new fact, `choco_temp_dir`, to discover
the appropriate temporary folder on disk in which to place the
7zip binary. It follows the pattern for using/testing facts as
the `choco_install_path` fact already used in this module, by
adding the logic for the fact in helper code.

This commit therefore also includes new tests to the module for
the helper code, the fact, the parameter, and the expected change
in behavior. It also includes updated documentation per the new
parameter.