(MODULES-6746) Remove Old Test Files
This change is to support MODULES-6746, which is a ticket to move this
module to the more standard method of testing with beaker-rspec. The
tests have already been moved to beaker-rspec, and this commit will
the older testing files that are no longer needed. A two step change was
used to keep CI green during the process of switching over.