(MAINT) Update chocolatey sources to artifactory
Prior to this commit the URIs for chocolatey feeds in the acceptance test
suite pointed to the old nexus feeds which have gone dark. This caused the
acceptance tests to fail.  This commit updates those URLs to point to the
new and functional artifactory feeds instead, fixing the acceptance tests.