(MODULES-6948) Fix collision in type generation
Prior to this commit running `puppet generate types` with the
Chocolatey module installed led to an erroneous error whilst
trying to build the type for `chocolateyfeature`. This was due
to a collision over the `features` method.

This commit replaces use of the word `features` in the provider
for `chocolateyfeature` with `choco_features`, ensures the rest
of the relevant code matches, and updates the tests.