(MODULES-4418) Add bypass_proxy to chocolateysource
Prior to this commit the `chocolateysource` type and provider did
not support the option to specify the `--bypass-proxy` flag when
managing a source. This option causes Chocolatey to ignore any
system proxies and was introduced in Chocolatey `0.10.4`.

This commit adds the `bypass_proxy` property to the type and
provider of `chocolateysource` to ensure that this property
can be configured on versions which support it.

This commit also updates the README documentation and CHANGELOG,
as well as providing new unit and acceptance tests and updating
any unit and acceptance tests which should include the new
setting.

As new features are added which depend on particular versions of
Chocolatey, it may be useful to refactor the tests and provider
to manage the addition of such features with less duplicated
code, but that is considered out-of-scope for this improvement.