(MODULES-5898) add admin_only to chocolateysource
Prior to this commit the `chocolateysource` type and provider did
not support the option to specify the `--admin-only` flag when
managing a source. This option causes Chocolatey to markthe source
as visible only to admins  and was introduced in Chocolatey `0.10.8`.

This commit adds the `admin_only` property to the type and
provider of `chocolateysource` to ensure that this property
can be configured on versions which support it.

This commit also updates the README documentation and CHANGELOG,
as well as providing new unit and acceptance tests and updating
any unit and acceptance tests which should include the new
setting.