(MODULES-5897) Add allow_self_service to chocolateysource
Prior to this commit the `chocolateysource` type and provider did
not support the option to specify the `--allow-self-service` flag
when managing a source. This option causes Chocolatey to mark the
source as being usable with the C4B Self-Service feature as of
C4B v1.10.0. It requires Chocolatey 0.10.4+ to set this option.

This commit adds the `allow_self_service` property to the type and
provider of `chocolateysource` to ensure that this property
can be configured on versions which support it.

This commit also updates the README documentation and CHANGELOG,
as well as providing new unit and acceptance tests and updating
any unit and acceptance tests which should include the new
setting.