(MODULES-7042) Document chocolatey_version
The existing documentation for the `chocolatey_version` parameter
of the `chocolatey` class is only in the init.pp file and briefly
mentioned in the documentation for the `chocolateyversion` fact.

None of these documentation sources cover that the parameter is
an optional informational flag to send to the class to help it
decide what behaviors to follow during execution.

A straight reading of the parameter name in the context of the
class is that it should pin Chocolatey to a particular version.
It does not and was not meant to do this. While a future ticket
may add this functionality it is important that we document and
clarify the current usage and make it easier for users to
understand.

This commit adds that documentation both to the class itself and
to the README.