(MODULES-8047) Fix puppet resource chocolateyconfig
Prior to this commit running `puppet resource chocolateyconfig` would
fail during the validation block which verified that no config items
were both :present *and* had an empty :value or were missing :value
altogether.

This is because Chocolatey represents a configuration item which is
using the defaults as an empty string. When collecting instances of
Chocolatey config items from the configuration file, it would find
the items, mark them as :present, and save them with a :value of "".

This would then fail the validation block.

During a `puppet apply` run the validation block is called before the
instances are fetched - and, therefore, before `@property_hash` is
populated. This allows us to distinguish between the two calls and error
on any declared resources which specify `:ensure` as `:present` but forget
to specify a `:value`.

This commit moves the validation logic from the type into the provider
in order to allow us to take advantage of the @property_hash information
during a `puppet resource` run.

This commit also updates the `get_config` method to treat any discovered
Chocolatey config item with an empty value as `:absent` - this matches
the behavior of `:ensure => :absent` which calls `choco config unset` -
this command replaces the existing value of a config item with an empty
string.

Now when returning the list of Chocolatey configs our model of them is
consistent, with `:absent` meaning that the configuration item is set
to the default.

This commit does not change the behavior of a `puppet apply` run.

This commit *does* include changes to the unit tests to reflect the
moving of the validation logic.