(MODULES-9224) Add no progress flag
Prior to this commit calls to Chocolatey to install or to
upgrade packages did not specify the `--no-progress` flag
even when it is available. This would sometimes cause the
puppet logs to include numerous lines of packages being
downloaded which is not particularly helpful.

This commit helps to reduce the noise in the logs by making
sure that the `--no-progress` flag is specified by default
if the version of Chocolatey on the machine supports it.

If, however, the resource also specifies the package setting
`verbose` as true then the `--no-progress` flag will not be
passed as the user is specifically requesting verbose output.

This commit includes updates to the provider, common helpers,
and spec tests. It does not include any additional user facing
documentation as this is an improvement which reduces noise
but does not otherwise require configuration.