(MODULES-9690) Redact Sensitive Commandline
This change redacts the commandline used with the `chocolateysource`
type if the user supplies a password.

If the user does not set a password for the source then no redaction
takes place. If a password is supplied then the entire commandline that
contains the password will be redacted.

This change also allows the user to specify a value to the `password`
parameter using the `Sensitive` data type, but note that this is not
required. The commandline will be redacted even if a normal string is
provided, and using the `Sensitive` type will cause Puppet to raise
a warning, even though nothing is wrong.