(maint) Replace chocolatey::params with hiera data
The chocolatey::params class in an internal detail with low value for
end users but cannot be made private for technical reasons so appears in
end-users documentation.  It is only useful to provide a default value
to $chocolatey::choco_install_location when the fact choco_install_path
is not defined, which can be achieved using hiera.