(GH-283) Fix deprecation error
In situations where the resource described in the manifest includes a
version in its ensure property and choco v1 or above is installed,
puppet runs would fail. This was happening of an issue where the
version of chocolatey was not being properly derived by choco_version.

In turn, would cause compiled_choco to return false causing the
module to default to using choco update rather than choco upgrade.

This commit fixes that by ensuring that the choco_version method
returns the correct puppet version.