(GH-285) Warn about secrets in install options
Prior to this commit, there was no guidance in the README that
warned users about the risks of passing secrets via install_options.

It looks like it existed at some point in the past but was removed
when we migrated to an automatically generated REFERENCE.md.

This commit adds the original warning back in to the README that was
introduced in GH-147.