(GH-297) Raise minimum Puppet required version
Prior to this commit, and according to issue #297, we introduced an
update on (1a3dc2c) that implemented a range feature that is only
possible to use with Puppet agent 6.15 or higher.

This commit aims to raise the minimum puppet agent version in the
metadata.json file to ensure that users comply with the minimum
requirements for the use of this feature.

This is a backwards incompatible change and will require a major version
bump.